Move blueprints to conductor
