+++ packages/core/assistant/package.json
