+++ packages/core/assistant/src/index.ts
@@ -6,4 +6,3 @@ export * from './extraction';
 export * from './research';
 export * from './session';
 export * from './types';
-export * from './blueprint';
