+++ packages/core/conductor/package.json
@@ -53,8 +53,11 @@
     "@dxos/protocols": "workspace:*",
     "@dxos/react-ui-kanban": "workspace:*",
     "@dxos/react-ui-table": "workspace:*",
+    "@dxos/schema": "workspace:*",
     "@dxos/util": "workspace:*",
     "@effect/platform": "0.80.12",
+    "@preact/signals-core": "^1.9.0",
+    "chalk": "^4.1.2",
     "effect": "3.14.21",
     "jsonpath-plus": "10.2.0",
     "ollama": "^0.5.11"
