+++ packages/core/conductor/src/blueprint/index.ts
@@ -5,3 +5,4 @@
 export * from './blueprint';
 export * from './machine';
 export * from './logger';
+export * from './compiler';
