+++ packages/core/conductor/tsconfig.json
@@ -33,6 +33,9 @@
       "path": "../../common/util"
     },
     {
+      "path": "../../sdk/schema"
+    },
+    {
       "path": "../../ui/react-ui-kanban"
     },
     {
